Qt6Uic: patch to not include the (generated) module config headers
qVersion() is defined in corelib/global/qlibraryinfo.cpp which checks various
Qt features and thus requires some module config headers -- which are
generated -- to be included. It also requires the generated source file(s)
qconfig-{develop,install}.cpp to be compiled.

In order to prevent complicating our build that much we manually inline
qVersion() instead. This seems reasonable because all qVersion() does is
return QT_VERSION_STR which is defined for every source file's compilation.

Furthermore, based on the upstream source code comments, the only difference
between qVersion() and QT_VERSION_STR is that the former is the runtime QtCore
version while the latter is the compile-time QtCore version. Given that we
compile QtCore into the Qt6Uic executable we need only the compile-time
version.