Include lower-level modules' config headers
In module config headers we include those of the lower-level modules because
there are a handful of cases where module config headers are included directly
instead of via the module global headers as is done in 95% of cases. This does
not cause any immediate problems upstream but it causes build failures for us
because we define some features conditionally upon lower-level modules'
features.

Direct inclusion of module config headers seems like a mistake based on a
survey of the Qt codebase but it doesn't look like upstream is willing to fix
these (by including the module global headers instead, as is done in 95% of
cases).